+++ src/module/system/rbac/service/role.service.ts
@@ -1,6 +1,7 @@
-import { Injectable } from '@nestjs/common';
-import { BaseService } from 'src/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
+import { BaseService, EStatus } from 'src/common';
 import { Transactional } from 'typeorm-transactional-cls-hooked';
+import { UserAbstractFacadeService } from '../../user/facade/user.facade.abstract';
 import {
   FindRoleReqBo,
   FindRoleResBo,
@@ -12,10 +13,15 @@ import {
   BatchUpdateReqBo,
 } from '../bo/role.bo';
 import { RoleAbstractRepoService } from '../repository/role.abstract';
+import { RbacSyncService } from './rbac-sync.service';
 
 @Injectable()
 export class RoleService extends BaseService {
-  constructor(private readonly roleRepoService: RoleAbstractRepoService) {
+  constructor(
+    private readonly roleRepoService: RoleAbstractRepoService,
+    private readonly userFacadeService: UserAbstractFacadeService,
+    private readonly rbacSyncService: RbacSyncService,
+  ) {
     super(RoleService.name);
   }
 
@@ -25,31 +31,46 @@ export class RoleService extends BaseService {
 
   async find(data: FindRoleReqBo): Promise<FindRoleResBo[]> {
     const result = await this.roleRepoService.find(data);
+    // TODO 获取角色管理权限包
     return result;
   }
 
   async findById(id: number): Promise<FindOneRoleResBo> {
     const result = await this.roleRepoService.findById(id);
+    // TODO 获取角色管理权限包
     return result;
   }
 
   async create(data: CreateRoleReqBo): Promise<RoleBo> {
     const result = await this.roleRepoService.create(data);
+    // TODO 设置角色管理权限包
     return result;
   }
 
   async updateById(id: number, data: UpdateRoleReqBo): Promise<void> {
-    // TODO 更新状态的后置条件 1、同步一下缓存
+    // TODO 设置角色管理权限包
     const result = await this.roleRepoService.updateById(
       id,
       this._.omit(data, []),
     );
+    // 更新状态的后置条件 1、同步一下缓存
+    if (data.status) {
+      const roles = await this.find({ status: EStatus.enable });
+      await this.rbacSyncService.sync(roles as RoleBo[]);
+    }
     return result;
   }
 
   async deleteById(id: number): Promise<void> {
-    // TODO 删除的前提条件 1、没有用户关联
-    // TODO 删除的后置条件 1、用户角色处理 2、同步一下缓存
+    // 删除的前提条件 1、没有用户关联
+    const users = await this.userFacadeService.findUidByRid(id);
+    if (users?.length) {
+      throw new BadRequestException('A1002');
+    }
+    // 删除的后置条件 1、用户角色处理 2、删除权限包关联关系 1、同步一下缓存
+    await this.roleRepoService.updateRolePgids(id, []);
+    const roles = await this.find({ status: EStatus.enable });
+    await this.rbacSyncService.sync(roles as RoleBo[]);
     return await this.roleRepoService.deleteById(id);
   }
